+++ datahub-web-react/src/app/search/AdvancedSearchAddFilterSelect.tsx
@@ -1,0 +1,50 @@
+import { Select } from 'antd';
+import * as React from 'react';
+import styled from 'styled-components';
+import { PlusOutlined } from '@ant-design/icons';
+
+import { FacetFilterInput } from '../../types.generated';
+import { FIELD_TO_LABEL } from './utils/constants';
+
+const StyledPlus = styled(PlusOutlined)`
+    margin-right: 6px;
+`;
+
+interface Props {
+    selectedFilters: Array<FacetFilterInput>;
+    onFilterFieldSelect: (value) => void;
+}
+
+const { Option } = Select;
+
+export const AdvancedSearchAddFilterSelect = ({ selectedFilters, onFilterFieldSelect }: Props) => {
+    return (
+        <Select
+            value={{
+                value: 'value',
+                label: (
+                    <div>
+                        <StyledPlus />
+                        Add Filter
+                    </div>
+                ),
+            }}
+            labelInValue
+            style={{ padding: 6, fontWeight: 500 }}
+            onChange={onFilterFieldSelect}
+            dropdownMatchSelectWidth={false}
+            filterOption={(_, option) => option?.value === 'null'}
+        >
+            {Object.keys(FIELD_TO_LABEL)
+                .sort((a, b) => FIELD_TO_LABEL[a].localeCompare(FIELD_TO_LABEL[b]))
+                .map((key) => (
+                    <Option
+                        disabled={key === 'entity' && !!selectedFilters.find((filter) => filter.field === 'entity')}
+                        value={key}
+                    >
+                        {FIELD_TO_LABEL[key]}
+                    </Option>
+                ))}
+        </Select>
+    );
+};
